feat: Allow routes to enforce a session by redirecting
The idea is that a page can show various links/RouteAnchors although
the user is not logged in, and once the use tries to navigate to them
(e.g. by clicking) the browser automatically starts the login flow.

This current issue is that the Redirect components already handle most
of this case. However, when the user logs out, without a page reloading,
the internal state is "not logged in, because logged off". This however
doesn't trigger a login, because the user chose to log off, or idled
out.

In order to make it easier to handle this, the redirect components now
act like the "authenticated" component, with the difference that they
would redirect, in case they are rendered, or trigger the log-in, in
the case they are created without an authenticated context.

BREAKING CHANGE: This change requires a few changes in the API.
Migration should be simple though.

Closes: #6