refactor: cleanups before the next release
Aside from a bunch of cleanups, this also makes the main configurations
non-exhaustive, in order to be prepared for future extensions.

BREAKING-CHANGE: In addition to the other configuration/option struct,
this also makes the main configuration non-exhaustive. So it is no
longer possible to directly create a new struct, but use the "new"
function instead and modify the struct using the `with_*` functions.